ENH: improved ACE instrument style
Improved the pysat compliance for the ACE instruments by assigning the satellite mission name to the platform and the instrument to the name.
